// src/MockData/Types.ts
export type TableData = {
	fullName: string;
	birthDate: string;
	address: string;
	gender: string;
	city: string;
	department: string;
	phone: string;
	maritalStatus: string;
	baptized: boolean;
	memberSince: string;
	churchRole: string;
	mentor: string;
};


export type OriginalProps = {
	fullName: string;
	birthDate: string;
	address: string;
	gender: string;
	city: string;
	department: string;
	phone: string;
	maritalStatus: string;
	baptized: boolean;
	memberSince: string;
	churchRole: string;
	mentor: string;
};